backend/requirements.txt
fastapi==0.111.0
uvicorn==0.24.0
httpx==0.26.0
python-dotenv==1.0.0
SQLAlchemy==2.0.22
psycopg2-binary==2.9.11
requests==2.31.0
python-jose[cryptography]==3.3.0
